Time: 97 ms (83.33%), Space: 16.8 MB (66.67%) - LeetHub

+++ 2518-number-of-great-partitions/2518-number-of-great-partitions.cpp
@@ -49,6 +49,6 @@ public:
             total%=mod;
         }
         long long value=find(nums,0,0,k);
-        return (int)(total-2*value+mod)%mod;
+        return (total-2*value+mod)%mod;
     }
 };
